+++ Systems/API/ZooExpoOrg.Api/Controllers/Expositions/ExpositionController.cs
@@ -2,13 +2,17 @@
 using AutoMapper;
 using Azure.Core;
 using Azure.Core.GeoJson;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
+using Newtonsoft.Json.Linq;
 using ZooExpoOrg.Api.Controllers.Clients;
 using ZooExpoOrg.Common.Exceptions;
+using ZooExpoOrg.Common.Security;
 using ZooExpoOrg.Services.Clients;
 using ZooExpoOrg.Services.Expositions;
 using ZooExpoOrg.Services.Logger;
+using ZooExpoOrg.Services.RightVerifier;
 
 namespace ZooExpoOrg.Api.Controllers.Expositions;
 
@@ -21,16 +25,19 @@ public class ExpositionController : Controller
     private readonly IAppLogger logger;
     private readonly IExpositionService expositionService;
     private readonly IMapper mapper;
+    private readonly IRightVerifierService rightVerifier;
 
     public ExpositionController(
         IAppLogger logger,
         IExpositionService expositionService, 
-        IMapper mapper
+        IMapper mapper,
+        IRightVerifierService rightVerifier
         )
     {
         this.logger = logger;
         this.expositionService = expositionService;
         this.mapper = mapper;
+        this.rightVerifier = rightVerifier;
     }
 
     [HttpGet("")]
@@ -56,10 +63,18 @@ public class ExpositionController : Controller
     }
 
     [HttpPost("")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> Create(CreateExpositionModel model)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfCreateExposition(jwtToken, model.OrganizerId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             var exposition = await expositionService.Create(model);
 
             return Ok(mapper.Map<PresintationExpositionModel>(exposition));
@@ -71,10 +86,18 @@ public class ExpositionController : Controller
     }
 
     [HttpPut("{id:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> Update(Guid id, UpdateExpositionModel model)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagExposition(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await expositionService.Update(id, model);
 
             return Ok();
@@ -86,10 +109,18 @@ public class ExpositionController : Controller
     }
 
     [HttpPut("{id:Guid}/subscribe/{clientId:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> Subscribe(Guid id, Guid clientId)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagClient(jwtToken, clientId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await expositionService.Subscribe(id, clientId);
 
             return Ok();
@@ -106,10 +137,21 @@ public class ExpositionController : Controller
     }
 
     [HttpPut("{id:Guid}/unsubscribe/{clientId:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> Unsubscribe(Guid id, Guid clientId)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagClient(jwtToken, clientId)))
+            {
+                if (!(await rightVerifier.VerifRightsOfManagExposition(jwtToken, id)))
+                {
+                    return BadRequest("Access denied.");
+                }
+            }
+
             await expositionService.Unsubscribe(id, clientId);
 
             return Ok();
@@ -121,10 +163,18 @@ public class ExpositionController : Controller
     }
 
     [HttpPut("{id:Guid}/participants/add/{animalId:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> AddParticipant(Guid id, Guid animalId)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagAnimal(jwtToken, animalId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await expositionService.AddParticipant(id, animalId);
 
             return Ok();
@@ -141,10 +191,21 @@ public class ExpositionController : Controller
     }
 
     [HttpPut("{id:Guid}/participants/delete/{animalId:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> DeleteParticipant(Guid id, Guid animalId)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagAnimal(jwtToken, animalId)))
+            {
+                if (!(await rightVerifier.VerifRightsOfManagExposition(jwtToken, id)))
+                {
+                    return BadRequest("Access denied.");
+                }
+            }
+
             await expositionService.DeleteParticipant(id, animalId);
 
             return Ok();
@@ -156,10 +217,18 @@ public class ExpositionController : Controller
     }
 
     [HttpDelete("{id:Guid}")]
+    [Authorize(AppScopes.UseScope)]
     public async Task<IActionResult> Delete(Guid id)
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagExposition(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await expositionService.Delete(id);
 
             return Ok();
